norminette

+++ src/starting_shell.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   starting_shell.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: macampos <mcamposmendes@gmail.com>         +#+  +:+       +#+        */
+/*   By: macampos <macampos@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/08 12:53:21 by macampos          #+#    #+#             */
-/*   Updated: 2024/08/23 12:52:08 by macampos         ###   ########.fr       */
+/*   Updated: 2024/08/23 17:19:50 by macampos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ void	child_process(char *user_input, char **envp, t_cmd *cmd, t_main *main)
 	}
 	if (check_builtins2(cmd, envp, main) == 1 && check[0] != -1)
 		not_builtin(check, envp, cmd, main);
-	check_builtins(cmd, envp, main, user_input);
+	main = check_builtins(cmd, envp, main, user_input);
 	free_every_thing(cmd, main, check);
 	exit(0);
 }
@@ -69,20 +69,14 @@ t_main	*execute_cmd(t_cmd *cmd, char **envp, t_main *main, char *user_input)
 	int	i;
 
 	i = 0;
-	main->pid = ft_calloc(10000, sizeof(pid_t));
 	main->cmd = ft_calloc(10000, sizeof(t_cmd));
 	if (cmd->next || check_builtins2(cmd, envp, main) == 1
 		|| cmd->redirectionoverall != 0)
 	{
+		main->pid = ft_calloc(10000, sizeof(pid_t));
 		while (cmd)
 		{
 			closepipes_helper(cmd);
-			if (!cmd->next && !ft_strncmp(cmd->args[0], "export", 6))
-				return (export(cmd, envp, main));
-			if (!cmd->next && !ft_strncmp(cmd->args[0], "unset", 5))
-				return (unset(cmd, main, envp));
-			if (!cmd->next && !ft_strncmp(cmd->args[0], "exit", 4))
-				return (exitt(cmd, envp, main));
 			cmd = execute_function_helper(main, i++, cmd, user_input);
 		}
 	}
@@ -98,18 +92,22 @@ t_main	*execute_function(char *user_input, char **envp,
 
 	main = execute_cmd(cmd, envp, main, user_input);
 	i = 0;
-	while (main->pid[i])
+	if (main->pid)
 	{
-		if (check_cmds(main) == 1 && !ft_strncmp(main->cmd[i]->args[0],
-				"cat", 3) && !main->cmd[i]->args[1])
+		while (main->pid[i])
 		{
-			getchar();
-			kill(main->pid[i], SIGUSR1);
+			if (check_cmds(main) == 1 && !ft_strncmp(main->cmd[i]->args[0],
+					"cat", 3) && !main->cmd[i]->args[1])
+			{
+				getchar();
+				kill(main->pid[i], SIGUSR1);
+			}
+			waitpid(main->pid[i], &main->status, 0);
+			i++;
 		}
-		waitpid(main->pid[i], &main->status, 0);
-		i++;
+		free(main->pid);
+		main->pid = NULL;
 	}
-	free(main->pid);
 	free(main->cmd);
 	if (WIFEXITED(main->status))
 		main->status = WEXITSTATUS(main->status);
